[FIX] Manual price rounding problem on pos order.
When on pos with manual_price_with_taxes active, an order can have price
rounding problem if a manual price is added on a product:
 * eg. 9,90 tax incl. -> 8,114754098(unit_price) + 1,785245902(22% tax).
When the order is created from pos, the unit price is saved rounded.
Then a differece between the payment(9,90) and the order total (9,89)
will be raised and the order status will remain to New instead of Payed.